Fixed doublebugs, added User Terminal functionality

+++ src/libnorad/cOrbitA.cc
@@ -39,7 +39,7 @@ cOrbitA::cOrbitA(std::string satNameA, int epochY, double epochD, double altitud
    int currentPlane = trunc(satIndex/satPerPlane);
    int planeIndex = (satIndex % (planes*satPerPlane))-(satPerPlane*currentPlane); //index of a satellite within a plane
    raan = ((360.0/planes)*currentPlane) * RADS_PER_DEG; //RAAN value, uniformly created so that there are equally spaced orbital planes for even coverage.
-   double phaseOffsetVal = ((360.0/satPerPlane)*(phaseOffset/planes))*currentPlane;
+   double phaseOffsetVal = ((360.0 / satPerPlane) * (static_cast<double>(phaseOffset) / planes)) * currentPlane;
    meanAnomaly = (((360.0/satPerPlane)*planeIndex))*RADS_PER_DEG; //Denotes the position of a satellite within its plane.
 
    m_jdEpoch = cJulian(epochYear, epochDay);
@@ -143,4 +143,3 @@ std::string cOrbitA::SatName(bool fAppendId /* = false */) const
 {
    return satName;
 }
-
